add pop up for install/uninstall
