+++ napari/_qt/widgets/qt_keyboard_settings.py
@@ -6,7 +6,6 @@ from qtpy.QtGui import QKeySequence
 from qtpy.QtWidgets import (
     QAbstractItemView,
     QComboBox,
-    QDialog,
     QHBoxLayout,
     QItemDelegate,
     QKeySequenceEdit,
@@ -25,7 +24,7 @@ from ...settings import get_settings
 from ...utils.action_manager import action_manager
 from ...utils.interactions import Shortcut
 from ...utils.translations import trans
-from ..qt_resources import get_stylesheet
+from ..widgets.qt_message_popup import WarnPopup
 
 # Dict used to format strings returned from converted key press events.
 # For example, the ShortcutTranslator returns 'Ctrl' instead of 'Control'.
@@ -468,7 +467,7 @@ class ShortcutEditor(QWidget):
         )
 
         # Create warning pop up and move it to desired position.
-        self._warn_dialog = KeyBindWarnPopup(
+        self._warn_dialog = WarnPopup(
             text=message,
         )
         self._warn_dialog.move(global_point)
@@ -500,41 +499,6 @@ class ShortcutEditor(QWidget):
             value[action_name] = list(shortcuts)
 
         return value
-
-
-class KeyBindWarnPopup(QDialog):
-    """Dialog to inform user that shortcut is already assigned."""
-
-    def __init__(
-        self,
-        parent=None,
-        text: str = "",
-    ):
-        super().__init__(parent)
-
-        self.setWindowFlags(Qt.FramelessWindowHint)
-
-        # Widgets
-        self._message = QLabel()
-        self._xbutton = QPushButton('x', self)
-        self._xbutton.setFixedSize(20, 20)
-
-        # Widget set up
-        self._message.setText(text)
-        self._message.setWordWrap(True)
-        self._xbutton.clicked.connect(self._close)
-        self._xbutton.setStyleSheet("background-color: rgba(0, 0, 0, 0);")
-
-        # Layout
-        main_layout = QVBoxLayout()
-        main_layout.addWidget(self._message)
-
-        self.setLayout(main_layout)
-
-        self.setStyleSheet(get_stylesheet(get_settings().appearance.theme))
-
-    def _close(self):
-        self.close()
 
 
 class ShortcutDelegate(QItemDelegate):
